+++ tools/daemon/list.do.pl
@@ -92,7 +92,7 @@ if( !GetOptions(
 }
 
 if( $ep->runner()->help()){
-	$ep->runner()->verbHelp( $defaults );
+	$ep->runner()->displayHelp( $defaults );
 	TTP::exit();
 }