+++ tools/dbms/backup.do.pl
@@ -152,7 +152,7 @@ if( !GetOptions(
 }
 
 if( $ep->runner()->help()){
-	$ep->runner()->verbHelp( $defaults );
+	$ep->runner()->displayHelp( $defaults );
 	TTP::exit();
 }