+++ tools/dbms/sql.do.pl
@@ -162,7 +162,7 @@ if( !GetOptions(
 }
 
 if( $ep->runner()->help()){
-	$ep->runner()->verbHelp( $defaults );
+	$ep->runner()->displayHelp( $defaults );
 	TTP::exit();
 }