+++ tools/dbms/telemetry.do.pl
@@ -260,7 +260,7 @@ if( !GetOptions(
 }
 
 if( $ep->runner()->help()){
-	$ep->runner()->verbHelp( $defaults );
+	$ep->runner()->displayHelp( $defaults );
 	TTP::exit();
 }